pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightspeed-sync"
version = "1.0.0"
description = "Lightspeed Retail (R-Series) + eCom (C-Series) API Updater"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lightspeed Sync Tool"}
]
keywords = ["lightspeed", "retail", "ecommerce", "api", "sync"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "tenacity>=8.2.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.0.0",
]

[project.scripts]
lightspeed-sync = "lightspeed_sync.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/lightspeed-sync"
Repository = "https://github.com/your-org/lightspeed-sync.git"
Issues = "https://github.com/your-org/lightspeed-sync/issues"

[tool.hatch.build.targets.wheel]
packages = ["lightspeed_sync"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
